% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt.R
\name{rt_get}
\alias{rt_get}
\title{get rt data}
\usage{
rt_get(region_code, dateym, service_key = NULL)
}
\arguments{
\item{region_code}{code}

\item{dateym}{year month}

\item{service_key}{api key}
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
get rt data
}
