% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtmolitr-package.R
\docType{package}
\name{rtmolitr-package}
\alias{rtmolitr}
\alias{rtmolitr-package}
\title{rtmolitr: 국토교통부 실거래가 api 사용 패키지}
\description{
국토교통부에서 재공하는 실거래가 정보를 가져오는 패키지입니다.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mrchypark/rtmolitr}
  \item Report bugs at \url{https://github.com/mrchypark/rtmolitr/issues}
}

}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (0000-0001-6474-2570)

}
\keyword{internal}
